pyproject.toml
[project]
name = "figma-bridge-webapp"
version = "0.1.0"
description = "Web app to extract Figma design data and convert to vibe coding format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.2",
]
